Implement MVC style, implement player data, implement changing of HP, implement primitive player UI
Unity does not really support "pure" MVC (Model, View, Controller) style, but it is still good practice to separate classes into separate scripts, which was done with this commit. In addition, the ability to set basic player data values such as HP and AP was implemented. To facilitate the displaying of those values, a primitive placeholder UI was added. Also added a list of outside assets used, which will be implemented into a credits screen later.