Move action/bullet properties into their own objects, separate actionmanager and currentmanager into controller and view, unify all private variables to start with underscore
Fully implementing the Model View Controller workflow, which results in the separating of changing Unity UI variables to Views and all other functions to Controllers. Since Actions and Bullets are their own objects, all their properties have been moved to be on the object itself to align better with OOP principles. Changed the Current grid to work as a true 2 dimensional array instead of having to have clearly defined variables for each column. Cleaned up code in general in preparation for implementing Actions.