Fix fatal bug when cancelling current space selection, add pattern deck and discard, add instantiation of cards to hand
There was a fatal bug where continuing within a while loop in a coroutine would also skip the yield return null that would force the coroutine to wait a frame. It is possible to just add another yield return null before continuing, but it seemed better to prefer not to use continue. Added ability to initialize a list of Pattern cards and add it to the deck, shuffle it, then draw from it and put it in the player hand. Functionality for discarding cards from hand.