Add phase progression to GameManager, prepare to rename all Current terms to Sight
Bullet is played with several phases similar to a card game; the basic functions of each phase has been added. This is a pre-commit to secure the functionality of the commit, as it was found that the project was using the wrong term for the place where bullets are placed on; it's Sight, not Current. Current is the "bag" that bullets are sent to from Center. This will be fixed to prevent further misunderstanding.