Add Esfir Volkova character, add Same number requirement checking, add Passive class
The first complete character has been added to the project, with their own passive, actions, and patterns. To implement this character, a new class has been added for character passives that is modular, able to be changed to whatever is needed. In addition, the requirement for same numbers in a pattern has been coded in. There is also now a GameMode enum that will be used later for game setup, but is currently built for the Score Attack mode.