Add Adelheid character, add facedown bullets, add diagonal search for bullets, add Intensity text
Added Adelheid, a character that relies on bullet colors, but can flip bullets face down to make them work as any color. To go along with this, support for facedown bullets was added, along with support for moving bullets diagonally through an action. Additionally, the current intensity is shown on screen, along with any extra bullets, so that it's possible to predict how many bullets will be sent next round.