Add Shield breaking mechanic to boss, add win/lose state for boss battles
Additionally, there is now a way to go back to the title screen once a game has been won/lose. Players can swap between the player and boss panel at will, the popup manager now just unactivates the popup instead of destroying and instantiating it repeatedly, and bullets are now sent to the boss's Incoming.