Add both promised bosses, add way to remove bullet as damage, whole game now feature complete
Both bosses as promised (Fugue in D Minor and The Fanged Sun) added, with all of their corresponding patterns, effects, passives, shields, etc. To go along with this, some new things have been added: The most important are observers to allow for passives that trigger at specific timings, and delegates to allow changing of data at specific timings. In addition, bullets moved past the bottom of the grid are now properly counted as damage.

As the game is now feature complete, the rest of the development time will be focused on adding in decent-looking UI and fixing game-breaking bugs by way of testing the game.