Final UI Revamp, final push before release
Added return to title screen button to main gameplay scene for easy resets. Added effects for cleared bullets and taking damage for better visual clarity. Moved swap panel button to a tidier location.

This was supposed to be the final August release. However many things came up near the end of the month that prevented me from working on the game until now. As the game is now currently in a decent prototype state, I am satisfied with the current state of the project.